Fixes for MP 1.0.1 Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
