+++ .github/workflows/build.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        java: [8]
+        java: [8, 11]
     name: build with jdk ${{matrix.java}}
 
     steps:
