+++ implementation/src/main/java/io/smallrye/graphql/bootstrap/datafetcher/AnnotatedPropertyDataFetcher.java
@@ -16,20 +16,11 @@
 
 package io.smallrye.graphql.bootstrap.datafetcher;
 
-import java.text.NumberFormat;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.TemporalAccessor;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Optional;
-
 import org.jboss.jandex.Type;
 
 import graphql.schema.DataFetchingEnvironment;
 import graphql.schema.PropertyDataFetcher;
 import io.smallrye.graphql.bootstrap.Annotations;
-import io.smallrye.graphql.bootstrap.schema.helper.CollectionHelper;
-import io.smallrye.graphql.bootstrap.schema.helper.FormatHelper;
 
 /**
  * Extending the default property data fetcher and take the annotations into account
@@ -37,74 +28,16 @@ import io.smallrye.graphql.bootstrap.schema.helper.FormatHelper;
  * @author Phillip Kruger (phillip.kruger@redhat.com)
  */
 public class AnnotatedPropertyDataFetcher extends PropertyDataFetcher {
-    private final FormatHelper formatHelper = new FormatHelper();
-
-    private DateTimeFormatter dateTimeFormatter = null;
-    private NumberFormat numberFormat = null;
-    private final CollectionHelper collectionHelper = new CollectionHelper();
+    private final TransformableDataFetcherHelper transformableDataFetcherHelper;
 
     public AnnotatedPropertyDataFetcher(String propertyName, Type type, Annotations annotations) {
         super(propertyName);
-        if (formatHelper.isDateLikeTypeOrCollectionThereOf(type)) {
-            this.dateTimeFormatter = formatHelper.getDateFormat(type, annotations);
-        }
-        if (formatHelper.isNumberLikeTypeOrCollectionThereOf(type)) {
-            this.numberFormat = formatHelper.getNumberFormat(annotations);
-        }
+        transformableDataFetcherHelper = new TransformableDataFetcherHelper(type, annotations);
     }
 
     @Override
     public Object get(DataFetchingEnvironment environment) {
         Object o = super.get(environment);
-        return get(o);
-    }
-
-    public Object get(Object o) {
-        if (Optional.class.isInstance(o)) {
-            Optional optional = Optional.class.cast(o);
-            if (optional.isPresent()) {
-                Object value = optional.get();
-                return Collections.singletonList(handleFormatting(value));
-            } else {
-                return Collections.emptyList();
-            }
-        } else if (Collection.class.isInstance(o)) {
-            Collection collection = Collection.class.cast(o);
-            Collection transformedCollection = collectionHelper.getCorrectCollectionType(o.getClass());
-            for (Object oo : collection) {
-                transformedCollection.add(get(oo));
-            }
-            return transformedCollection;
-        } else {
-            return handleFormatting(o);
-        }
-    }
-
-    private Object handleFormatting(Object o) {
-        if (dateTimeFormatter != null) {
-            return handleDateFormatting(o);
-        } else if (numberFormat != null) {
-            return handleNumberFormatting(o);
-        } else {
-            return o;
-        }
-    }
-
-    private Object handleDateFormatting(Object o) {
-        if (TemporalAccessor.class.isInstance(o)) {
-            TemporalAccessor temporalAccessor = (TemporalAccessor) o;
-            return dateTimeFormatter.format(temporalAccessor);
-        } else {
-            return o;
-        }
-    }
-
-    private Object handleNumberFormatting(Object o) {
-        if (Number.class.isInstance(o)) {
-            Number number = (Number) o;
-            return numberFormat.format(number);
-        } else {
-            return o;
-        }
+        return transformableDataFetcherHelper.transform(o);
     }
 }
