+++ pom.xml
@@ -17,7 +17,7 @@
 
     <properties>
         <version.eclipse.microprofile.config>1.4</version.eclipse.microprofile.config>
-        <version.eclipse.microprofile.graphql>1.1-SNAPSHOT</version.eclipse.microprofile.graphql>
+        <version.eclipse.microprofile.graphql>1.0.1</version.eclipse.microprofile.graphql>
         <sonar.coverage.jacoco.xmlReportPaths>target/site/jacoco/jacoco.xml</sonar.coverage.jacoco.xmlReportPaths>
     </properties>
 
@@ -82,11 +82,6 @@
             <dependency>
                 <groupId>io.smallrye</groupId>
                 <artifactId>smallrye-graphql-servlet</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>io.smallrye</groupId>
-                <artifactId>smallrye-graphql-vertx</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
